+++ packages/article/__tests__/ios/article-scaling.ios.test.js
@@ -1,0 +1,3 @@
+import scalingTests from "../scaling.native";
+
+scalingTests();