+++ packages/icons/src/icons/prop-types.js
@@ -4,9 +4,9 @@ import PropTypes from "prop-types";
 // 1) if width and height is not set Svg scales to fill the parent container
 // 2) the svgs package creates different markup when title or strokecolour is set
 export default {
-  title: PropTypes.string,
-  strokeColour: PropTypes.string,
   fillColour: PropTypes.string,
   height: PropTypes.number,
+  strokeColour: PropTypes.string,
+  title: PropTypes.string,
   width: PropTypes.number
 };