+++ packages/markup-forest/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/newsuk/times-components/compare/@times-components/markup-forest@1.0.1...@times-components/markup-forest@1.0.2) (2018-08-13)
+
+**Note:** Version bump only for package @times-components/markup-forest
+
+
+
+
+
 <a name="1.0.1"></a>
 ## [1.0.1](https://github.com/newsuk/times-components/compare/@times-components/markup-forest@1.0.0...@times-components/markup-forest@1.0.1) (2018-08-09)