+++ packages/tealium-utils/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.4"></a>
+## [0.6.4](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.6.3...@times-components/tealium-utils@0.6.4) (2018-08-13)
+
+**Note:** Version bump only for package @times-components/tealium-utils
+
+
+
+
+
 <a name="0.6.3"></a>
 ## [0.6.3](https://github.com/newsuk/times-components/compare/@times-components/tealium-utils@0.6.2...@times-components/tealium-utils@0.6.3) (2018-08-09)